requirements.txt
Pillow==9.3.0
python-magic==0.4.27
tqdm==4.64.1
